# apps/web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "alexgetman.com API",
    "description": "API endpoints for alexgetman.com including JSON/RSS feeds and simple pageview tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://alexgetman.com"
    }
  ],
  "paths": {
...
    "/stats/pageview": {
      "post": {
        "summary": "Record Pageview",
        "description": "Sends a pageview event for tracking analytics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "example": "/"
                  }
                },
                "required": ["path"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully tracked"
          }
        }
      }
    },
...
  }
}